sched: Remove blocked load contribution of dying tasks
Task being dequeued for the last time (state == TASK_DEAD) are dequeued
with the DEQUEUE_SLEEP flag which causes their load contribution to be
added to the runqueue blocked load. Hence the rq blocked load will
contain load that is gone away. The issue only exists for the root
cfs_rq as cgroup_exit() doesn't set DEQUEUE_SLEEP for task group exits.

If runnable+blocked load is to be used as a better estimate for cpu
load the dead task contributions need to be removed to prevent
load_balance() (idle_balance() in particular) from over-estimating the
cpu load.