sched: Add capacity_curr_max to the runqueue.
In order to make the scheduler aware of the maximum capacity
that a given CPU can be scaled to under capping resulting from
a thermal or cpufreq event. We store the capacity_curr_max on
the runqueue. The capacity_orig_of() returns capacity_curr_max
in order to accommodate for the dynamic changes in the maximum
capacity that a CPU can be scaled to.

Signed-off-by: Kapileshwar Singh <kapileshwar.singh@arm.com>