[HMP] Bugfix potential hardlock on hmp_idle_pull
In softirqd, when all the softirq have been completed, a call to
__schedule() could lead to an hmp_idle_pull. The HMP idle pull code makes
use of a timer to avoid the big CPUs to enter a long latency idle state
while a big task is going to be pushed.

It is quite unlikely but it could happens that "hrtimer_start" calls
"wakeup_softirqd" which in turn calls "try_to_wake_up" on the current task,
i.e. the ksoftirqd for the big CPU doing an idle pull.
In this case, the scheduler in the try_to_wake_up will busy loop with
preempt/irq disabled waiting for a never happening dequeue of
the current task.

This patch fix the unlikely but technically possible problem using a
slightly different API call for the "keepalive" timer.
Since this timer is use just to constraint the CPUIdle framework on the
selection of the next possible idle state, this patch enqueue the timer
using a call which ensures not to generate a wakeup. This could cause the
missing of the scheduled timer, which is however not an issue given the
specific role of the keepalive timer.

Reported-by: Chao Xie <cxie4@marvell.com>
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Reviewed-by: Chris Redpath <chris.redpath@arm.com>