sched: System-wide sched domain
Currently we have the problem that we can't express system wide energy by
attaching energy data onto a sched group (sg) which itself is attached to
a sched domain (sd). The scheduler does not allow sd's with only one sg
since this is useless from the scheduling perspective.

This patch introduces the possibility to have such a system-wide sd. The
only purpose of this sd is to hold a single sg on which we can attach
system-wide energy data in form of sched_group::sched_group_energy.

The new sd topology flag SD_SHARE_ENERGY is necessary to indicate to
sd_init() that it has to treat this sd in a special way by zeroing it out
and only set sd::flags to SD_SHARE_ENERGY and sd::name to the name
provided via the arch specific sched_domain_topology_level table
(normally SYS).
Furthermore we use it in CFS's rebalance_domains() to bail out early,
even before the decay of the newidle max times.

build_sched_domains() won't normally allow to have two adjacent sd's with
a span of cpu_map because it would bail out of the first for_each_cpu loop
allowing only one sd with a span of cpu_map.
Since we want this additional SYS sd on top of the MC sd of a single
cluster system or on top of the DIE sd of a multi-cluster system (both of
them already spanning the whole cpu_map) this patch deletes this if
condition.

All non EAS related occurrences of for_each_domain() (except the former
mentioned one in rebalance_domains() and the one in show_schedstat()) are
not effected because they either check for SD_LOAD_BALANCE or
SD_WAKE_AFFINE. There is obviously the cost of an additional iteration
when SYS sd is enabled.

Tested on ARM TC2 (2 cluster (Cortex A7/A15 - big.LITTLE) and Pandaboard
(dual core (single cluster) Cortex A9).

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>