sched/fair: rename signals update methods to match data structures
The (static) methods used to manipulate the load and utilization signals
are using different naming conventions which does not match with the
specific data structure which is going to be updated.

This patch rename all the setter methods to be consistent with the specific
signal and data structure being updated.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>