sched/fair: rename RQ "load" and "utilization" signals
The "load" and "utilization" signals are used to track respectively
the average time a SE spent in RUNNABLE (on RQ) and RUNNING (on CPU)
state.

This patch rename these signals to:
- make them better integrated with the container data structure
- use a more consistent naming convention
- simplify code readability by using shorter yet still meaningful names

The new set of signals introduced by this patch is a couple of name
consistent pairs:
  <signal> and <signal>_blocked
where <signal> is "load" and "utilization".

Each pair represents the "on-RQ" and "off-RQ" contribution of a SE on
the corresponding <signal>.

Documentation is updated to better describe the meaning of each signal.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>