sched/fair: add a field to energy_env to keep track of group max usage
Functions called by energy_aware_wake_cpu() make use of an energy_env
struct to give some context to inner helpers functions.

As we are going to need some of the calculations inner functions do
(namely the maximum usage of a certain scheduling group), add a field
to the energy_env struct that can be used to save the calculated max
usage.

Signed-off-by: Juri Lelli <juri.lelli@arm.com>