sched/fair: disambiguate sched_group_energy function from struct
The symbol sched_group_energy is used both for a struct and a function.
Even if that's completely legitimate, it seems to me we could improve
documentation and code readability by renaming the function to better
distinguish the two concepts.

For example, "estimate_sg_energy" IMHO could better represent its goal:
to compute an estimation of the energy consumption provided an energy
model and the scheduling modification defined by the energy_env
description.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>