FIXUP: ASoC: mediatek: add missing clocks
[Problem]
There are some clocks required for audio but not ultilized in CCF yet.
Before the new CCF version is available, we can only operate these clocks
directly by audio driver for now.

[Solution]
Add clock related registers, and set these registers directly. A debugfs
is also added to check the correctness of register settings.
Once the CCF is updated with the missing clocks, this FIXUP can be removed.

Signed-off-by: Koro Chen <koro.chen@mediatek.com>

BUG=chrome-os-partner:38746
TEST=build and boot oak-rev0

Change-Id: I95fd55cd532f3a8fa15053ea570c10b5f3145b89