[GTK] Avoid computations when moving small tasks
This is done here just to easy the integration of the code computing the
capability difference. However, it seems to make sense to avoid in general
the computation of the energy diff function when moving really small
tasks. Perhaps, these tasks should just be moved to a LITTLE CPU, perhaps
the first one, which could be easily determined or defined at boot time
as the CPU with the highest energy efficiency.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>