sched/cpufreq_sched_cfs: implement direct API
Instead of using get_cpu_usage() we can let each CPU
request the capacity it needs. The gov's kthread is
responsible for aggregating requests.

A benefit of this new API is shown in task_tick_fair(),
where we can request a transition to max opp only when
really needed.

Signed-off-by: Juri Lelli <juri.lelli@arm.com>

Conflicts:
	kernel/sched/fair.c
	kernel/sched/sched.h

Conflicts:
	kernel/sched/fair.c