CHROMIUM: config: mediatek - add DRM_POWERVR_ROGUE_MT8173E1
This CONFIG option was added for drm/img-rogue 1.6@3539569

Having to select an individual version of the GPU as a CONFIG option is
going the wrong way and making the kernel driver even harder to use.

But, unfortunately we are stuck with this until Imagination implements
runtime GPU version/feature detection.

 echo 'CONFIG_DRM_POWERVR_ROGUE_MT8173E1=y' \
   >> chromeos/config/arm64/chromiumos-mediatek.flavour.config
 chromeos/scripts/kernelconfig oldconfig

Signed-off-by: Rufus Hamade <rufus.hamade@imgtec.com>

BUG=chrome-os-partner:39263
TEST=Boot Oak board to Freon UI, login and browse BBC website.

Change-Id: Ic2e4155246a79c9ab7b798a86a05413a72463feb