HACK: CHROMIUM: drm/img-rogue: bypass BVNC 'V' check
In general, the drm/img-rogue driver is not able to support different BVNC
versions at runtime; support for a particular BVNC must be hard coded at
build time.  To enforce this, the driver fails to load if an incompatible
BVNC is found.

However, this it is not practical to require different kernel CONFIG
options for different GPU variants.

In particular, different spins of the mt8173 have different BVNC values:
mt8173 E1 has BVNC=4.29.2.51
mt8173 E2 has BVNC=4.40.2.51

So, disable the "V" check for now, until proper 'runtime version/feature'
support is implemented.

Signed-off-by: Cawa Cheng <cawa.cheng@mediatek.com>
Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:40578
TEST=test on oak-rev0 and oak-rev1 board

Change-Id: I7c50042d5db4db933838f123a35f823640962e10