sched: export get_cpu_usage & capacity_orig_of
get_cpu_usage and capacity_orig_of are useful for a cpu frequency
scaling policy which is based on cfs load tracking and cpu capacity
metrics. Expose these calls in sched.h so that they can be used in such
a policy.

Signed-off-by: Michael Turquette <mturquette@linaro.org>

Conflicts:
	kernel/sched/fair.c

Fix return type of get_cpu_usage() (this changed from int in
EASv3 to unsinged long in EASv4).