sched: Add over-utilization/tipping point indicator
Energy-aware scheduling can only operate when the system is not
overutilized. There must be cpu time available to place tasks based on
utilization in an energy-aware fashion, i.e. to pack tasks on
energy-efficient cpus without creating additional latency conditions
for those tasks.

In case the system operates above this tipping point the tasks have to
be placed based on task and cpu load in the classical way of spreading
tasks across as many cpus as possible to avoid these latency conditions
as much as possible.

The point in which a system switches from being not overutilized to
being overutilzed is called the tipping point.

Such a tipping point indicator on a sched domain (sd) as the system
boundary is introduced here. As soon as one cpu of a sd is overutilized
the whole sd is declared overutilized as well. A cpu becomes
overutilized when its utilization is higher that 80% of its capacity.

The implementation takes advantage of the shared sd that is common
across all the sds at a level. The new flag introduced is placed in
this structure so that all the sd the same level share the flag.

Load balancing is skipped in case the energy model is present and the sd
is not overutilized because under this condition the predominantly
load-per-capacity driven load-balancer should not interfere with the
energy-aware wakeup placement based on utilization.

In case the total utilization of a sd is greater than the total sd
capacity the flag is set at the parent sd level to let other sched
groups (sg) help getting rid of the overutilzation of cpus.

This implementation has corner scenarios. Consider an sg with n cpus and
n+1 70% tasks. This would be a case for load balance to happen in a
parent sd. But the total group utilization is not high enough for the
load balance to be triggered in the parent sd. But this could be
considered as a academic scenario, as during task wake up these tasks
will be placed more appropriately.

Signed-off-by: Thara Gopinath <thara.gopinath@linaro.org>
Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>