blmp: add load tracking signal
This foundation patch adds the load tracking signals required by HMP as
they was defined in previous LSK kernel version 3.10.

The main signal added is the "load_avg_ratio" of each task, which is a
variant of "load_avg_contrib" not scaled by the task priority.
It is calculated like:

  se->avg.load_avg_ratio =
    runnable_avg_sum * NICE_0_LOAD / (runnable_avg_period + 1)

The sum of "load_avg_ratio" of each task into a RQ is accumulated by:

  rq_of(cfs_rq)->avg.load_avg_ratio

which is updated each time a task is added/removed to/from the RQ.

Signed-off-by: Paul Turner <pjt@google.com>
Reviewed-by: Ben Segall <bsegall@google.com>
Signed-off-by: Morten Rasmussen <Morten.Rasmussen@arm.com>
Signed-off-by: Chris Redpath <chris.redpath@arm.com>
Signed-off-by: Liviu Dudau <liviu.dudau@arm.com>
Signed-off-by: Jon Medhurst <tixy@linaro.org>