blmp: migration strategy: FORK
Goals:
1. spawn new tasks on big cores, but
2. keep kernel threads and init task direct childs on the slower domain

System services are generally started by init, whilst kernel threads are
started by kthreadd. We do not want to give those tasks a head start, as
this costs power for very little benefit. We do however wish to do that
for tasks which the user launches.

Further, some tasks allocate per-cpu timers directly after launch which
can lead to those tasks being always scheduled on a big CPU when there
is no computational need to do so. Not promoting services to big CPUs on
launch will prevent that unless a service allocates their per-cpu
resources after a period of intense computation, which is not a common
pattern.

Signed-off-by: Chris Redpath <chris.redpath@arm.com>
Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>
Signed-off-by: Jon Medhurst <tixy@linaro.org>
Signed-off-by: Liviu Dudau <Liviu.Dudau@arm.com>
Signed-off-by: Robin Randhawa <robin.randhawa@arm.com>