blmp: migration strategy: FORCED
Goals:
1. periodically check for the load ratio of the running tasks
2. move running tasks with a significative load ratio to a
   CPU of the faster domain
3. move running tasks with a limited load ratio to a CPU
   of the slower domain
4. do not have more than one task per CPUs on the faster domain

Force migration of runnable tasks when load balancing is triggered.

Tasks running in a less capable hmp_domain may change to become more
demanding and should therefore be migrated up. Since they are unlikely
go through the select_task_rq_fair() path anytime soon they need special
attention.

This patch introduces a period check (SCHED_TICK) of the haviest
runnable task on all runqueues and sets up a forced migration using
stop_machine_no_wait() if the task needs to be migrated.

Ideally, this should not be implemented by polling all runqueues.

* hmp_force_up_migration
  Checks RQs for tasks that need to be migrated to a faster cpu.
  This code is serialzed by the "hmp_force_migration" spinlock, which
  ensure its execution by a single CPU, and it could be triggered
  by the scheduler tick (if needed) and on nohz idle balancing.

.:: Up migration to idle CPU

When a forced up-migration should move a task to an idle CPU, instead of
stopping the task to be migrated until the target CPU becomes available,
we interrupt the target CPU and ask it to pull a task.  This lets the
current eligible task continue executing on the original CPU while the
target CPU wakes.
A pinned timer is used to prevent the pulling CPU going back into
power-down with pending up-migrations.

Signed-off-by: Chris Redpath <chris.redpath@arm.com>
Signed-off-by: Jon Medhurst <tixy@linaro.org>
Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>
Signed-off-by: Morten Rasmussen <Morten.Rasmussen@arm.com>