blmp: experimental: Frequency Invariant (FI) scale
Use load as a representation the used CPU "potential compute capacity"
rather than a representation its "current compute capacity".

If CPUFreq is enabled, scale load in accordance with current frequency.
As long as the CPUFreq subsystem correctly reports the current operating
frequency, the scaling should self tune.

Powersave/performance CPUFreq governors are detected and scaling is
disabled while these governors are in use.  Indeed,  when a
single-frequency governor is in use, potential CPU capacity is static.

An additional sysfs attribute is exposed:

  /sys/kernel/hmp/frequency_invariant_load_scale

which allows to enable (write 1) or disable (write 0) this
(experimental) feature.

Signed-off-by: Chris Redpath <chris.redpath@arm.com>