CHROMIUM: arm64: dts: mt8173-oak: add sbs-battery to i2c_tunnel
Add a smart battery node on the i2c_tunnel through which the AP talks to
the EC.

There is a cros-ec-sbs snippet in arch/arm, however, that snippet is not
upstream, and needs further per-board configuration.  It is pretty small
too, so it is more convenient to just include the node directly in the
board .dts file.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:38746
TEST=build and boot oak-rev0

Change-Id: If80422f363557e20dfcdafa78fbaa4d8637e3bb9