[GTK] sched: initial support for EAS tuning via CGroups
NOTE: this is just an initial description of the commit.

The current implementation of the EAS scheduler has been designed to
exploit an energy model in order to support an energy efficient allocation
of tasks on available CPUs. Such a design, for the time being, does not
allow to boost some tasks for higher performances if required at the expense
of a comfortable increase in energy consumption.

To this purpose, the EAS scheduler wants to provide a "knob" which allows
to tune how much the system is going to be optimised for energy efficiency
vs performances.

This patch provides a simple CGroups based interface to define per
task-group EAS tuning knobs. For the time being:
- just one "margin" tunable is exposed, which can be configured in
  range [0..100]
- no constraints are enforced on "margin" values between parent groups

Change-Id: Ic538e9a9ef4def3ec44befbfe68cdfd5c2f60b7d
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>