WIP: sched/cpufreq_sched_cfs: add the get_expected_capacity function
The expected capacity for a CPU is computed by SchedDVFS code using either
get_cpu_usage() or task_utilization() in different code paths.
This approach seems to be incomplete since we think that the selection of
an OPP should always be based on the actual utilisation of the CPU, i.e. by
considering the contribution of all the tasks currently "active" on it.

This patch provides a consolidation of the approach used to compute the
expected capacity of a CPU which accomplish two main goals:
1. always use get_cpu_usage() to get an updated view on the current CPU
   usage (i.e. active + blocked utilisation)
2. setup the ground for the integration of utilisation boosting once the
   SchedTune margin is used

The first gaol has required to migrate some SchedDVFS paths from the usage
of task_utilization() to get_cpu_usage(). Specifically:
  a) dequeue_task_fair
     which is expected to work properly even when using that signal
  b) select_task_rq_fair
     which is expected to be probably broken, at least in the case of a
     target CPU which a long time sleeper where the blocked utilisation
     is not yet properly decayed once we are on that function

NOTE: this last issue is expected to be better verified an fixed in a following
      patch while still we believe that the final solution should be based
      on the usage active+blocked utilisation signals.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>