WIP: SchedTune: add nrg/cap gain selection attributes
Boosting the utilisation of a task could imply the selection of a CPU/OPP
and cluster which has an higher energy impact. In that conditions the
energy_diff() function will discard the scheduling proposal under evaluation
thus impacting the possibility to boost the performances of a task.
A trade-off between increased energy consumption and performances benefits
should be evaluated.

This patch introduces the support to define a couple of thresholds in the
Performance-Energy variation Space, which allows to identify the:
- Performance Boost (B) region
  top right quadrant: both nrg and perf are positive
- Performance Constraining (C) region
  bottom-left quadrant: both nrg and perf are negative

This first implementation binds both the thresholds values to the single
value of the boost margin, thus allowing to use a single knob to configure
all the thresholds.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>