WIP: SchedTune: add energy_payoff computation
Boosting the utilisation of a task could imply the selection of a CPU/OPP
and cluster which has an higher energy impact. In that conditions the
energy_diff() function will discard the scheduling proposal under evaluation
thus impacting the possibility to boost the performances of a task.
A trade-off between increased energy consumption and performances benefits
should be evaluated.

This patch introduces the required support to evaluate the "energy payoff"
corresponding to a certain variation on expected task performances.
The proposed scheduling decision is described in term of expected variation
of energy and performance. These values are used to verify in which region
of the Performance-Energy Space the solution becomes and thus returns
an energy_payoff metrics which is positive if
- the increased energy is compensated by a configured performance gain
  i.e. either in Optimal (O) or Performance Boost (B) region
- the decreased energy is compliant with a configured performance constraint
  i.e. in Performance Constraint (C) region

Otherwise the returned energy_payoff will be negative, i.e. not in the above
cases or in the Suboptimal (S) region.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>