WIP: sched: add support to get the boosted task utilisation
The utilisation signal of a task could be boosted by a configurable margin
in order to make it appears as an higher demanding workload on some
specific code paths.
This patch provides the get_boosted_task_utilization() function which
could be used in all the code paths where looking at a (eventually) boosted
utilization could be used to bias the behaviours of the scheduler.
The main customers of this interface will be:
- SchedDVFS, for the selection of a (possibly) higher OPP
- EAS (wakeup path), for the selection of a (possibly) higher capacity CPU

NOTE: the boosting of the actual task utilisation is enabled only while
      the scheduler is working in EAS mode, i.e. the root domain where
      the task is currently scheduled is not marked as being
      "overutilized".

Indeed, the task boosting support is disabled at run-time when the system
is saturated because in such a system benefits of boosting becomes
negligible (e.g. we are already running at the highest OPP).

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>