mali: make setting up the OPPs a weak function
Setting up the OPPs via a module relies on the module being loaded at
the right time.  For Juno that means that it has to be loaded after the
SCPI clocks are loaded but before the core mali driver.  Getting that
ordering right is quite a pain and now that the SCPI is a platform
driver, using module_init is a guarantee for failure, as it's always
going to happen before any platform driver is loaded.

Instead, make setting up the opps a function that the driver
provides as a weak, dummy function that can be overridden by platforms
providing a strong function if they need to set up OPPs.

Signed-off-by: Javi Merino <javi.merino@arm.com>
Signed-off-by: Jon Medhurst <tixy@linaro.org>
[Fix merge conflicts]
Signed-off-by: Arvind Chauhan <arvind.chauhan@arm.com>