WIP: arm64: topology: Define OAK energy and link to SDs
This is basically same hack we have in ARM for TC2/Juno were the energy
model tables are hard-coded in topology.c.

NOTE: this patch provides Energy Model (EM) data for a JUNO board.
As soon as an EM for OAK has been generated/verified by Mediatek we will
update again that patch.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>

BUG=chromium:48849
TEST=build and boot tested

Change-Id: I79f85715d0d1ca04295dbd3e7d735e46de484dca