WIP: DEBUG: add get_boosted_task_utilization tracepoint
A tracepoint reporting the task boosted utilization is generated each time
we call the get_boosted_task_utilization. This "on demand" approach should
help on reducing the number of events generated while still getting an
updated view of the boosted utilization in all the decision points where
we need to get a choice based on that value.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>